global exception handling for backend API

+++ todo/src/main/java/spring/todo/Category/CategoryService.java
@@ -2,6 +2,8 @@ package spring.todo.Category;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import spring.todo.exceptions.ResourceNotFoundException;
+import spring.todo.exceptions.DuplicateTaskException;
 
 import java.util.List;
 
@@ -16,14 +18,20 @@ public class CategoryService {
     }
 
     public Category createCategory(Category category) {
+        if (categoryRepository.existsByNameIgnoreCase(category.getName())) {
+            throw new DuplicateTaskException("Category name '" + category.getName() + "' already exists.");
+        }
         return categoryRepository.save(category);
     }
 
     public Category getCategoryById(Long id) {
-        return categoryRepository.findById(id).orElse(null);
+        return categoryRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Category not found with id: " + id));
     }
 
     public void deleteCategory(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Category not found with id: " + id);
+        }
         categoryRepository.deleteById(id);
     }
 }
